(#1900) User IDs are now longs instead of ints

+++ WebApp/src/uk/ac/exeter/QuinCe/jobs/JobManager.java
@@ -261,7 +261,7 @@ public class JobManager {
     long addedID = DatabaseUtils.NO_DATABASE_RECORD;
 
     // Get the user's ID
-    int ownerID = NO_OWNER;
+    long ownerID = NO_OWNER;
 
     if (null != owner) {
       // Check that the user exists
@@ -307,7 +307,7 @@ public class JobManager {
         if (NO_OWNER == ownerID) {
           stmt.setNull(1, java.sql.Types.INTEGER);
         } else {
-          stmt.setInt(1, ownerID);
+          stmt.setLong(1, ownerID);
         }
 
         stmt.setTimestamp(2, new Timestamp(System.currentTimeMillis()));
